# Hotfix avoid entering a bogus equation loop through recursion