--- a/library/Booster/Pattern/Rewrite.hs
+++ b/library/Booster/Pattern/Rewrite.hs
@@ -665,10 +665,12 @@ performRewrite doTracing def mLlvmLibrary mSolver mbMaxDepth cutLabels terminalL
                     rewriteTrace $ RewriteSimplified traces (Just r)
                     pure $ Just p
                 Left r@(EquationLoop (t : ts)) -> do
-                    let termDiffs = zipWith (curry mkDiffTerms) (t : ts) ts
                     logError "Equation evaluation loop"
-                    logSimplify $
-                        "produced the evaluation loop: " <> Text.unlines (map (prettyText . fst) termDiffs)
+                    logOtherNS "booster" (LevelOther "ErrorDetails") $
+                        let termDiffs = zipWith (curry mkDiffTerms) (t : ts) ts
+                            l = length ts
+                         in "Evaluation loop of length " <> prettyText l <> ": \n" <>
+                              Text.unlines (map (prettyText . fst) termDiffs)
                     rewriteTrace $ RewriteSimplified traces (Just r)
                     pure $ Just p
                 Left other -> do