Yesterday's change in #430  to include a top-level `notBool` (so that simplifications can be applied to it) ends up recursively calling evaluation of the same term (i.e., a bogus evaluation loop of size 1) in some KEVM test proofs.
This fix adds back the recursion breaking flag to protect against this (the test behaves normally again with this fix).