@jberthold 2023-12-19
Comparison against master from Friday last week:

| Test                                     | 7-HOTFIX-regression-entering-bogus-equation-loop time | 3-master-6e29375 time | (7-HOTFIX-regression-entering-bogus-equation-loop/3-master-6e29375) time 
|------------------------------------------|-------------------------------------------------------|-----------------------|--------------------------------------------------------------------------
| benchmarks/overflow00-overflow-spec.k    | 46.48                                                 | 65.95                 | 0.7047763457164518                                                       
| mcd/cat-file-addr-pass-rough-spec.k      | 62.4                                                  | 79.51                 | 0.784806942522953                                                        
| benchmarks/staticloop00-a0lt10-spec.k    | 55.28                                                 | 69.12                 | 0.7997685185185185                                                       
| mcd/vat-deny-diff-fail-rough-spec.k      | 119.3                                                 | 136.1                 | 0.8765613519470977                                                       
| erc20/ds/transferFrom-failure-1-c-spec.k | 123.7                                                 | 138.66                | 0.8921101976056541                                                       
| erc20/hkg/transfer-success-2-spec.k      | 51.32                                                 | 54.94                 | 0.9341099381143065                                                       
| mcd/flipper-tau-pass-spec.k              | 79.9                                                  | 85.46                 | 0.9349403229581091                                                       
| erc20/ds/transferFrom-success-2-spec.k   | 97.76                                                 | 104.34                | 0.9369369369369369                                                       
| mcd/vat-move-diff-rough-spec.k           | 140.39                                                | 148.29                | 0.9467260098455729                                                       
| bihu/forwardToHotWallet-success-2-spec.k | 150.19                                                | 156.81                | 0.9577833046361839                                                       
| benchmarks/ecrecover00-sigvalid-spec.k   | 95.4                                                  | 99.34                 | 0.9603382323334004                                                       
| mcd/flopper-cage-pass-spec.k             | 62.46                                                 | 64.91                 | 0.9622554305962102                                                       
| erc20/hkg/transferFrom-success-2-spec.k  | 146.2                                                 | 141.11                | 1.0360711501665365                                                       
| bihu/forwardToHotWallet-failure-1-spec.k | 71.32                                                 | 68.12                 | 1.04697592483852                                                         
| mcd/functional-spec.k                    | 480.12                                                | 455.75                | 1.0534722984092155                                                       
| erc20/hkg/approve-spec.k                 | 54.73                                                 | 51.09                 | 1.0712468193384224                                                       
| erc20/ds/approve-success-spec.k          | 67.04                                                 | 62.46                 | 1.0733269292347103                                                       
| erc20/hkg/transferFrom-success-1-spec.k  | 74.88                                                 | 69.26                 | 1.0811435171816344                                                       
| mcd/cat-exhaustiveness-spec.k            | 170.92                                                | 157.69                | 1.0838987887627622                                                       
| erc20/ds/transferFrom-failure-1-d-spec.k | 100.77                                                | 86.36                 | 1.1668596572487262                                                       
| benchmarks/storagevar00-spec.k           | 62.98                                                 | 48.47                 | 1.2993604291314214                                                       
| benchmarks/requires01-a0gt0-spec.k       | 67.1                                                  | 50.95                 | 1.3169774288518152                                                       
| TOTAL                                    | 2380.64                                               | 2394.6899999999996    | 0.9941328522689785